Implement move impossible and custom input
Changed the result of having an impossible move. Previously, it would skip turns. Now, the opponent collects all their own korgools and then the kazans are compared. Whoever has more korgools in their kazan wins. 

Implemented custom input handling, so that the game can be set to a particular starting point and both the Board and GUI will correctly reflect this change.